+++ whitecanvas/backend/mock/canvas.py
@@ -109,7 +109,7 @@ class Canvas:
 
 @protocols.check_protocol(protocols.CanvasGridProtocol)
 class CanvasGrid:
-    def __init__(self, heights: list[int], widths: list[int], app: str = "default"):
+    def __init__(self, heights: list[float], widths: list[float], app: str = "default"):
         self._background_color = np.array([1, 1, 1, 1], dtype=np.float32)
         self._figsize = (100, 100)
