+++ whitecanvas/backend/plotly/canvas.py
@@ -200,7 +200,7 @@ class Canvas:
 
 @protocols.check_protocol(protocols.CanvasGridProtocol)
 class CanvasGrid:
-    def __init__(self, heights: list[int], widths: list[int], app: str = "default"):
+    def __init__(self, heights: list[float], widths: list[float], app: str = "default"):
         from plotly.subplots import make_subplots
 
         if app == "notebook":
